moved event stream from eventstore to repository
