+++ repository.go
@@ -11,7 +11,6 @@ import (
 type eventStore interface {
 	Save(events []Event) error
 	Get(id string, aggregateType string, afterVersion Version) ([]Event, error)
-	EventStream() observer.Stream
 }
 
 type snapshotStore interface {
@@ -33,6 +32,8 @@ type aggregate interface {
 type Repository struct {
 	eventStore eventStore
 	snapshotStore snapshotStore
+	events observer.Property  // A property to which all event changes for all event types are published
+
 }
 
 // NewRepository factory function
@@ -40,6 +41,7 @@ func NewRepository(eventStore eventStore, snapshotStore snapshotStore) *Reposito
 	return &Repository{
 		eventStore: eventStore,
 		snapshotStore: snapshotStore,
+		events: observer.NewProperty(nil),
 	}
 }
 
@@ -49,6 +51,11 @@ func (r *Repository) Save(aggregate aggregate) error {
 	if err != nil {
 		return err
 	}
+	// publish the saved events to the events stream
+	for _, event := range aggregate.changes() {
+		r.events.Update(event)
+	}
+	// aggregate are saved to the event store now its safe to update the internal aggregate state
 	aggregate.updateVersion()
 	return nil
 }
@@ -74,7 +81,7 @@ func (r *Repository) Get(id string, aggregate aggregate) error {
 	return nil
 }
 
-// EventStream returns a stream with all saved events
+// EventStream returns a stream where all saved event will be published
 func (r *Repository) EventStream() observer.Stream {
-	return r.eventStore.EventStream()
+	return r.events.Observe()
 }
